requirements.txt
functions-framework==3.5.0
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.100.0 